norgesgruppen/pyproject.toml
[project]
name = "nm_ai_image"
version = "0.0.1"
description = "A short description of the project."
authors = [{ name = "DIG" }]
readme = "README.md"
requires-python = "~=3.10.0"
dependencies = [
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "lightning>=2.6.0",
    "timm>=1.0.0",
    "albumentations>=1.4.0",
    "torchmetrics>=1.0.0",
    "segmentation-models-pytorch>=0.3.0",
    "Pillow>=10.0.0",
    "grad-cam>=1.5.0",
    "cmaes>=0.11.0",
    "optuna>=4.0.0",
    "optuna-integration[pytorch_lightning]>=4.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "mlflow>=2.10.0",
    "pymoo>=0.6.0",
    "SALib>=1.5.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "websockets>=12.0",
    "ultralytics>=8.4.24",
    "fpdf2>=2.8.7",
    "onnx>=1.20.1",
    "onnxruntime>=1.23.2",
    "onnxslim>=0.1.89",
    "ensemble-boxes>=1.0.9",
    "streamlit>=1.55.0",
    "onnxconverter-common>=1.16.0",
]

[project.optional-dependencies]
azure = [
    "azure-ai-ml>=1.12.0",
    "azure-identity>=1.15.0",
]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["nm_ai_image*"]

[tool.ruff]
line-length = 120
src = ["nm_ai_image"]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.lint.isort]
known-first-party = ["nm_ai_image"]
force-sort-within-sections = true

[tool.pytest.ini_options]
testpaths = ["test"]
